first draft of revised create-item ecosystem
