+++ docs/modtools/create-item.rst
@@ -14,30 +14,30 @@ Usage
 
     modtools/create-item <options>
 
+Examples
+--------
+
+``modtools/create-item -u 23145 -i WEAPON:ITEM_WEAPON_PICK -m INORGANIC:IRON -q4``
+    Have unit 23145 create an exceptionally crafted iron pick.
+``modtools/create-item -u 323 -i CORPSEPIECE:NONE -m CREATURE_MAT:DWARF:BRAIN``
+    Have unit 323 produce a lump of brain.
+``modtools/create-item -i BOULDER:NONE -m INORGANIC:ADAMANTINE -c 5``
+``modtools/create-item -i BOULDER:NONE -m PLANT_MAT:MUSHROOM_HELMET_PLUMP:DRINK``
+
 Options
 -------
 
-    -creator id
-        specify the id of the unit who will create the item,
-        or \\LAST to indicate the unit with id df.global.unit_next_id-1
-        examples:
-            0
-            2
-            \\LAST
-    -material matstring
-        specify the material of the item to be created
-        examples:
-            INORGANIC:IRON
-            CREATURE_MAT:DWARF:BRAIN
-            PLANT_MAT:MUSHROOM_HELMET_PLUMP:DRINK
-    -item itemstr
-        specify the itemdef of the item to be created
-        examples:
-            WEAPON:ITEM_WEAPON_PICK
-    -quality qualitystr
-        specify the quality level of the item to be created (df.item_quality)
-        examples: Ordinary, WellCrafted, FinelyCrafted, Masterful, or 0-5
-    -matchingShoes
-        create two of this item
-    -matchingGloves
-        create two of this item, and set handedness appropriately
+``-u``, ``--unit <id>`` (default: first citizen)
+    The ID of the unit to use as the item's creator. You can also pass the
+    string "\\LAST" to use the most recently created unit.
+``-i``, ``--item <itemdef>`` (required)
+    The def string of the item you want to create.
+``-m``, ``--material`` (required)
+    That def string of the material you want the item to be made out of.
+``-q``, ``--quality`` (default: ``0``, which is ``df.item_quality.Ordinary``)
+    The quality of the created item.
+``-d``, ``--description`` (required if you are creating a slab)
+    The text that will be engraved on the created slab.
+``-c``, ``--count`` (default: ``1``)
+    The number of items to create. If the item is stackable, this will be the
+    stack size.
