+++ gui/create-item.lua
@@ -1,54 +1,20 @@
 -- A gui-based item creation script.
 -- author Putnam
 -- edited by expwnent
---@module = true
 
+local argparse = require('argparse')
+local createitem = reqscript('modtools/create-item')
 local eventful = require('plugins.eventful')
-local guidm = require('gui.dwarfmode')
 local script = require('gui.script')
 local utils = require('utils')
 
-local no_quality_item_types = utils.invert{
-    'BAR',
-    'SMALLGEM',
-    'BLOCKS',
-    'ROUGH',
-    'BOULDER',
-    'WOOD',
-    'CORPSE',
-    'CORPSEPIECE',
-    'REMAINS',
-    'MEAT',
-    'FISH',
-    'FISH_RAW',
-    'VERMIN',
-    'PET',
-    'SEEDS',
-    'PLANT',
-    'SKIN_TANNED',
-    'PLANT_GROWTH',
-    'THREAD',
-    'DRINK',
-    'POWDER_MISC',
-    'CHEESE',
-    'FOOD',
-    'LIQUID_MISC',
-    'COIN',
-    'GLOB',
-    'ROCK',
-    'EGG',
-    'BRANCH',
-}
-
-------------------------------
-
-function getGenderString(gender)
+local function getGenderString(gender)
     local sym = df.pronoun_type.attrs[gender].symbol
     if not sym then return '' end
     return '(' .. sym .. ')'
 end
 
-function getCreatureList()
+local function getCreatureList()
     local crList = {}
     for k,cr in ipairs(df.global.world.raws.creatures.alphabetic) do
         for kk,ca in ipairs(cr.caste) do
@@ -60,7 +26,7 @@ function getCreatureList()
     return crList
 end
 
-function getCreaturePartList(creatureID, casteID)
+local function getCreaturePartList(creatureID, casteID)
     local crpList = {{'generic'}}
     for k,crp in ipairs(df.global.world.raws.creatures.all[creatureID].caste[casteID].body_info.body_parts) do
         local str = crp.name_singular[0][0]
@@ -69,7 +35,7 @@ function getCreaturePartList(creatureID, casteID)
     return crpList
 end
 
-function getCreaturePartLayerList(creatureID, casteID, partID)
+local function getCreaturePartLayerList(creatureID, casteID, partID)
     local crplList = {{'whole'}}
     for k,crpl in ipairs(df.global.world.raws.creatures.all[creatureID].caste[casteID].body_info.body_parts[partID].layers) do
         local str = crpl.layer_name
@@ -78,7 +44,7 @@ function getCreaturePartLayerList(creatureID, casteID, partID)
     return crplList
 end
 
-function getCreatureMaterialList(creatureID, casteID)
+local function getCreatureMaterialList(creatureID, casteID)
     local crmList = {}
     for k,crm in ipairs(df.global.world.raws.creatures.all[creatureID].material) do
         local str = crm.id
@@ -87,7 +53,12 @@ function getCreatureMaterialList(creatureID, casteID)
     return crmList
 end
 
-function getRestrictiveMatFilter(itemType, opts)
+local function getCreatureRaceAndCaste(caste)
+    return df.global.world.raws.creatures.list_creature[caste.index],
+            df.global.world.raws.creatures.list_caste[caste.index]
+end
+
+local function getRestrictiveMatFilter(itemType, opts)
     if opts.unrestricted then return nil end
     local itemTypes = {
         WEAPON = function(mat, parent, typ, idx)
@@ -125,7 +96,7 @@ function getRestrictiveMatFilter(itemType, opts)
     return itemTypes[df.item_type[itemType]]
 end
 
-function getMatFilter(itemtype, opts)
+local function getMatFilter(itemtype, opts)
     local itemTypes = {
         SEEDS = function(mat, parent, typ, idx)
             return mat.flags.SEED_MAT
@@ -167,36 +138,7 @@ function getMatFilter(itemtype, opts)
     return itemTypes[df.item_type[itemtype]] or getRestrictiveMatFilter(itemtype, opts)
 end
 
-function createItem(mat, itemType, quality, creator, description, amount)
-    local item = df.item.find(dfhack.items.createItem(itemType[1], itemType[2], mat[1], mat[2], creator))
-    local item2 = nil
-    assert(item, 'failed to create item')
-    quality = math.max(0, math.min(5, quality - 1))
-    item:setQuality(quality)
-    if df.item_type[itemType[1]] == 'SLAB' then
-        item.description = description
-    end
-    if df.item_type[itemType[1]] == 'GLOVES' then
-        --create matching gloves
-        item:setGloveHandedness(1)
-        item2 = df.item.find(dfhack.items.createItem(itemType[1], itemType[2], mat[1], mat[2], creator))
-        assert(item2, 'failed to create item')
-        item2:setQuality(quality)
-        item2:setGloveHandedness(2)
-    end
-    if df.item_type[itemType[1]] == 'SHOES' then
-        --create matching shoes
-        item2 = df.item.find(dfhack.items.createItem(itemType[1], itemType[2], mat[1], mat[2], creator))
-        assert(item2, 'failed to create item')
-        item2:setQuality(quality)
-    end
-    if tonumber(amount) > 1 then
-        item:setStackSize(amount)
-        if item2 then item2:setStackSize(amount) end
-    end
-end
-
-function qualityTable()
+local function qualityTable()
     return {{'None'},
         {'-Well-crafted-'},
         {'+Finely-crafted+'},
@@ -206,7 +148,7 @@ function qualityTable()
     }
 end
 
-function showItemPrompt(text, item_filter, hide_none)
+local function showItemPrompt(text, item_filter, hide_none)
     require('gui.materials').ItemTypeDialog{
         prompt = text,
         item_filter = item_filter,
@@ -215,11 +157,10 @@ function showItemPrompt(text, item_filter, hide_none)
         on_cancel = script.mkresume(false),
         on_close = script.qresume(nil),
     }:show()
-
     return script.wait()
 end
 
-function showMaterialPrompt(title, prompt, filter, inorganic, creature, plant) --the one included with DFHack doesn't have a filter or the inorganic, creature, plant things available
+local function showMaterialPrompt(title, prompt, filter, inorganic, creature, plant)
     require('gui.materials').MaterialDialog{
         frame_title = title,
         prompt = prompt,
@@ -231,346 +172,90 @@ function showMaterialPrompt(title, prompt, filter, inorganic, creature, plant) -
         on_cancel = script.mkresume(false),
         on_close = script.qresume(nil),
     }:show()
-
     return script.wait()
 end
 
-function usesCreature(itemtype)
-    typesThatUseCreatures = {
-        REMAINS = true,
-        FISH = true,
-        FISH_RAW = true,
-        VERMIN = true,
-        PET = true,
-        EGG = true,
-        CORPSE = true,
-        CORPSEPIECE = true,
-    }
-    return typesThatUseCreatures[df.item_type[itemtype]]
-end
-
-local function getCreatureRaceAndCaste(caste)
-    return df.global.world.raws.creatures.list_creature[caste.index],
-            df.global.world.raws.creatures.list_caste[caste.index]
-end
-
-local CORPSE_PIECES = utils.invert{'BONE', 'SKIN', 'CARTILAGE', 'TOOTH', 'NERVE', 'NAIL', 'HORN', 'HOOF', 'CHITIN',
-    'SHELL', 'IVORY', 'SCALE'}
-local HAIR_PIECES = utils.invert{'HAIR', 'EYEBROW', 'EYELASH', 'MOUSTACHE', 'CHIN_WHISKERS', 'SIDEBURNS'}
-local LIQUID_PIECES = utils.invert{'BLOOD', 'PUS', 'VENOM', 'SWEAT', 'TEARS', 'SPIT', 'MILK'}
-
--- this part was written by four rabbits in a trenchcoat (ppaawwll)
-function createCorpsePiece(creator, bodypart, partlayer, creatureID, casteID, generic)
-    -- (partlayer is also used to determine the material if we're spawning a "generic" body part (i'm just lazy lol))
-    creatureID = tonumber(creatureID)
-    -- get the actual raws of the target creature
-    local creatorRaceRaw = df.creature_raw.find(creatureID)
-    local wholePart = false
-    casteID = tonumber(casteID)
-    bodypart = tonumber(bodypart)
-    partlayer = tonumber(partlayer)
-    -- somewhat similar to the bodypart variable below, a value of -1 here means that the user wants to spawn a whole body part. we set the partlayer to 0 (outermost) because the specific layer isn't important, and we're spawning them all anyway. if it's a generic corpsepiece we ignore it, as it gets added to anyway below (we can't do it below because between here and there there's lines that reference the part layer
-    if partlayer == -1 and not generic then
-        partlayer = 0
-        wholePart = true
-    end
-    -- get body info for easy reference
-    local creatorBody = creatorRaceRaw.caste[casteID].body_info
-    local layerName
-    local layerMat = 'BONE'
-    local tissueID
-    local liquid = false
-    -- in the hackWish function, the bodypart variable is initialized to -1, which isn't changed if the spawned item is a corpse
-    local isCorpse = bodypart == -1 and not generic
-    if not generic and not isCorpse then -- if we have a specified body part and layer, figure all the stuff out about that
-        -- store the tissue id of the specific layer we selected
-        tissueID = tonumber(creatorBody.body_parts[bodypart].layers[partlayer].tissue_id)
-        local mats = {}
-        -- get the material name from the material itself
-        for i in string.gmatch(dfhack.matinfo.getToken(creatorRaceRaw.tissue[tissueID].mat_type, creatureID), '([^:]+)') do
-            table.insert(mats, i)
-        end
-        layerMat = mats[3]
-        layerName = creatorBody.body_parts[bodypart].layers[partlayer].layer_name
-    elseif not isCorpse then -- otherwise, figure out the mat name from the dual-use partlayer argument
-        -- no "whole" option at the start of the generic creature material selection prompt means that the value we get is actually further along than intended
-        partlayer = partlayer + 1
-        layerMat = creatorRaceRaw.material[partlayer].id
-        layerName = layerMat
-    end
-    -- default is MEAT, so if anything else fails to change it to something else, we know that the body layer is a meat item
-    local item_type = 'MEAT'
-    -- get race name and layer name, both for finding the item material, and the latter for determining the corpsepiece flags to set
-    local raceName = string.upper(creatorRaceRaw.creature_id)
-    -- every key is a valid non-hair corpsepiece, so if we try to index a key that's not on the table, we don't have a non-hair corpsepiece
-    -- we do the same as above but with hair
-    -- if the layer is fat, spawn a glob of fat and DON'T check for other layer types
-    if layerName == 'FAT' then
-        item_type = 'GLOB'
-    elseif CORPSE_PIECES[layerName] or HAIR_PIECES[layerName] then -- check if hair
-        item_type = 'CORPSEPIECE'
-    elseif LIQUID_PIECES[layerName] then
-        item_type = 'LIQUID_MISC'
-        liquid = true
-    end
-    if isCorpse then
-        item_type = 'CORPSE'
-        generic = true
-    end
-    local itemType = dfhack.items.findType(item_type .. ':NONE')
-    local itemSubtype = dfhack.items.findSubtype(item_type .. ':NONE')
-    local material = 'CREATURE_MAT:' .. raceName .. ':' .. layerMat
-    local materialInfo = dfhack.matinfo.find(material)
-    local item_id = dfhack.items.createItem(itemType, itemSubtype, materialInfo['type'], materialInfo.index, creator)
-    local item = df.item.find(item_id)
-    if liquid then
-        local bucketMat = dfhack.matinfo.find('PLANT_MAT:NETHER_CAP:WOOD')
-        if not bucketMat then
-            for i,n in ipairs(df.global.world.raws.plants.all) do
-                if n.flags.TREE then
-                    bucketMat = dfhack.matinfo.find('PLANT_MAT:' .. n.id .. ':WOOD')
-                end
-                if bucketMat then break end
-            end
-        end
-        local prevCursorPos = guidm.getCursorPos()
-        local bucketType = dfhack.items.findType('BUCKET:NONE')
-        local bucket = df.item.find(dfhack.items.createItem(bucketType, -1, bucketMat.type, bucketMat.index, creator))
-        dfhack.items.moveToContainer(item, bucket)
-        guidm.setCursorPos(creator.pos)
-        dfhack.run_command('spotclean')
-        guidm.setCursorPos(prevCursorPos)
-    end
-
-    -- if the item type is a corpsepiece, we know we have one, and then go on to set the appropriate flags
-    if item_type == 'CORPSEPIECE' then
-        if layerName == 'BONE' then -- check if bones
-            item.corpse_flags.bone = true
-            item.material_amount.Bone = 1
-        elseif layerName == 'SKIN' then -- check if skin/leather
-            item.corpse_flags.leather = true
-            item.material_amount.Leather = 1
-            -- elseif layerName == "CARTILAGE" then -- check if cartilage (NO SPECIAL FLAGS)
-        elseif layerName == 'HAIR' then -- check if hair (simplified from before)
-            item.corpse_flags.hair_wool = true
-            item.material_amount.HairWool = 1
-            if materialInfo.material.flags.YARN then
-                item.corpse_flags.yarn = true
-                item.material_amount.Yarn = 1
-            end
-        elseif layerName == 'TOOTH' or layerName == 'IVORY' then -- check if tooth
-            item.corpse_flags.tooth = true
-            item.material_amount.Tooth = 1
-        elseif layerName == 'NERVE' then    -- check if nervous tissue
-            item.corpse_flags.skull1 = true -- apparently "skull1" is supposed to be named "rots/can_rot"
-            item.corpse_flags.separated_part = true
-            -- elseif layerName == "NAIL" then -- check if nail (NO SPECIAL FLAGS)
-        elseif layerName == 'HORN' or layerName == 'HOOF' then -- check if nail
-            item.corpse_flags.horn = true
-            item.material_amount.Horn = 1
-        elseif layerName == 'SHELL' then
-            item.corpse_flags.shell = true
-            item.material_amount.Shell = 1
-        end
-        -- checking for skull
-        if not generic and not isCorpse and creatorBody.body_parts[bodypart].token == 'SKULL' then
-            item.corpse_flags.skull2 = true
-        end
-    end
-    local matType
-    -- figure out which material type the material is (probably a better way of doing this but whatever)
-    for i in pairs(creatorRaceRaw.tissue) do
-        if creatorRaceRaw.tissue[i].tissue_material_str[1] == layerMat then
-            matType = creatorRaceRaw.tissue[i].mat_type
-        end
-    end
-    if item_type == 'CORPSEPIECE' or item_type == 'CORPSE' then
-        --referencing the source unit for, material, relation purposes???
-        item.race = creatureID
-        item.normal_race = creatureID
-        item.normal_caste = casteID
-        -- usually the first two castes are for the creature's sex, so we set the item's sex to the caste if both the creature has one and it's a valid sex id (0 or 1)
-        if casteID < 2 and #(creatorRaceRaw.caste) > 1 then
-            item.sex = casteID
-        else
-            item.sex = -1 -- it
-        end
-        -- on a dwarf tissue index 3 (bone) is 22, but this is not always the case for all creatures, so we get the mat_type of index 3 instead
-        -- here we also set the actual referenced creature material of the corpsepiece
-        item.bone1.mat_type = matType
-        item.bone1.mat_index = creatureID
-        item.bone2.mat_type = matType
-        item.bone2.mat_index = creatureID
-        -- skin (and presumably other parts) use body part modifiers for size or amount
-        for i = 0,200 do                          -- fuck it this works
-            -- inserts
-            item.body.bp_modifiers:insert('#', 1) --jus,t, set a lot of it to one who cares
-        end
-        -- copy target creature's relsizes to the item's's body relsizes thing
-        for i,n in pairs(creatorBody.body_parts) do
-            -- inserts
-            item.body.body_part_relsize:insert('#', n.relsize)
-            item.body.components.body_part_status:insert(i, creator.body.components.body_part_status[0]) --copy the status of the creator's first part to every body_part_status of the desired creature
-            item.body.components.body_part_status[i].missing = true
-        end
-        for i in pairs(creatorBody.layer_part) do
-            -- inserts
-            item.body.components.layer_status:insert(i, creator.body.components.layer_status[0]) --copy the layer status of the creator's first layer to every layer_status of the desired creature
-            item.body.components.layer_status[i].gone = true
-        end
-        if item_type == 'CORPSE' then
-            item.corpse_flags.unbutchered = true
-        end
-        if not generic then
-            -- keeps the body part that the user selected to spawn the item from
-            item.body.components.body_part_status[bodypart].missing = false
-            -- restores the selected layer of the selected body part
-            item.body.components.layer_status[creatorBody.body_parts[bodypart].layers[partlayer].layer_id].gone = false
-        elseif generic then
-            for i in pairs(creatorBody.body_parts) do
-                for n in pairs(creatorBody.body_parts[i].layers) do
-                    if item_type == 'CORPSE' then
-                        item.body.components.body_part_status[i].missing = false
-                        item.body.components.layer_status[creatorBody.body_parts[i].layers[n].layer_id].gone = false
-                    else
-                        -- search through the target creature's body parts and bring back every one which has the desired material
-                        if creatorRaceRaw.tissue[creatorBody.body_parts[i].layers[n].tissue_id].tissue_material_str[1] == layerMat and creatorBody.body_parts[i].token ~= 'SKULL' and not creatorBody.body_parts[i].flags.SMALL then
-                            item.body.components.body_part_status[i].missing = false
-                            item.body.components.layer_status[creatorBody.body_parts[i].layers[n].layer_id].gone = false
-                            -- save the index of the bone layer to a variable
-                        end
-                    end
-                end
-            end
-        end
-        -- brings back every tissue layer associated with a body part if we're spawning the entire thing
-        if wholePart then
-            for i in pairs(creatorBody.body_parts[bodypart].layers) do
-                item.body.components.layer_status[creatorBody.body_parts[bodypart].layers[i].layer_id].gone = false
-                item.corpse_flags.separated_part = true
-                item.corpse_flags.unbutchered = true
-            end
-        end
-        -- DO THIS LAST or else the game crashes for some reason
-        item.caste = casteID
-    end
-end
-
-function hackWish(unit, opts)
-    script.start(function()
-        local matok, mattype, matindex
-        local partlayerok, partlayerID = false, 0
-        local qualityok, quality = false, df.item_quality.Ordinary
-        local itemok, itemtype, itemsubtype = showItemPrompt('What item do you want?',
+local default_accessors = {
+    get_unit = function(opts)
+        return tonumber(opts.unit) and df.unit.find(tonumber(opts.unit)) or
+                dfhack.gui.getSelectedUnit(true)
+    end,
+    get_item_type = function()
+        return showItemPrompt('What item do you want?',
             function(itype) return df.item_type[itype] ~= 'FOOD' end, true)
-        local corpsepieceGeneric
-        local bodypart = -1
-        if not itemok then return end
-        if not usesCreature(itemtype) then
-            matok, mattype, matindex = showMaterialPrompt('Wish', 'And what material should it be made of?',
-                not opts.unrestricted and getMatFilter(itemtype, opts) or nil)
-            if not matok then return end
-        else
-            local creatureok, _, creatureTable = script.showListPrompt('Wish', 'What creature should it be?',
-                COLOR_LIGHTGREEN, getCreatureList(), 1, true)
-            if not creatureok then return end
-            mattype, matindex = getCreatureRaceAndCaste(creatureTable[3])
-        end
-        if df.item_type[itemtype] == 'CORPSEPIECE' then
-            local bodpartok, bodypartLocal = script.showListPrompt('Wish', 'What body part should it be?',
-                COLOR_LIGHTGREEN,
-                getCreaturePartList(mattype, matindex), 1, true)
-            -- createCorpsePiece() references the bodypart variable so it can't be local to here
-            bodypart = bodypartLocal
-            if bodypart == 1 then
-                corpsepieceGeneric = true
-            end
-            if not bodpartok then return end
-            if not corpsepieceGeneric then -- probably a better way of doing this tbh
-                partlayerok, partlayerID = script.showListPrompt('Wish', 'What tissue layer should it be?',
-                    COLOR_LIGHTGREEN,
-                    getCreaturePartLayerList(mattype, matindex, bodypart - 2), 1, true)
-            else
-                partlayerok, partlayerID = script.showListPrompt('Wish', 'What creature material should it be?',
-                    COLOR_LIGHTGREEN,
-                    getCreatureMaterialList(mattype, matindex), 1, true)
-            end
-            if not partlayerok then return end
-        elseif not no_quality_item_types[df.item_type[itemtype]] then
-            qualityok, quality = script.showListPrompt('Wish', 'What quality should it be?', COLOR_LIGHTGREEN,
-                qualityTable())
-            if not qualityok then return end
-        end
-        local description
-        if df.item_type[itemtype] == 'SLAB' then
-            local descriptionok
-            descriptionok, description = script.showInputPrompt('Slab', 'What should the slab say?', COLOR_WHITE)
-            if not descriptionok then return end
-        end
-        if bodypart ~= -1 then
-            --the offsets here are cause indexes in lua are wonky (some start at 0, some start at 1), so we adjust for that, as well as the index offset created by inserting the "generic" option at the start of the body part selection prompt
-            bodypart = bodypart - 2
-        end
-        local count = opts.count
-        if not count then
-            repeat
-                local amountok, amount = script.showInputPrompt('Wish', 'How many do you want? (numbers only!)',
-                    COLOR_LIGHTGREEN)
-                if not amountok then return end
-                count = tonumber(amount)
-            until count
-        end
-        if not mattype or not itemtype then
-            return false
-        end
-        if df.item_type.attrs[itemtype].is_stackable then
-            createItem({mattype, matindex}, {itemtype, itemsubtype}, quality, unit, description, count)
-        else
-            for _ = 1,count do
-                if itemtype == df.item_type.CORPSEPIECE or itemtype == df.item_type.CORPSE then
-                    createCorpsePiece(unit, bodypart, partlayerID - 2, mattype, matindex, corpsepieceGeneric)
-                else
-                    createItem({mattype, matindex}, {itemtype, itemsubtype}, quality, unit, description, 1)
-                end
-            end
-        end
-        return true
-    end)
-end
-
-local function get_first_citizen()
-    local citizens = dfhack.units.getCitizens()
-    if not citizens or not citizens[1] then
-        qerror('Could not choose a creator unit. Please select one in the UI')
-    end
-    return citizens[1]
-end
-
-local function get_creator(opts)
-    return tonumber(opts.unit) and df.unit.find(tonumber(opts.unit)) or
-            dfhack.gui.getSelectedUnit(true) or
-            get_first_citizen()
-end
-
-if dfhack_flags.module then
-    return
-end
-
-validArgs = utils.invert{
-    'startup',
-    'unrestricted',
-    'unit',
-    'count',
+    end,
+    get_mat = function(itype, opts)
+        return showMaterialPrompt('Wish', 'And what material should it be made of?',
+            not opts.unrestricted and getMatFilter(itype, opts) or nil)
+    end,
+    get_creature_mat = function()
+        local creatureok, _, creatureTable = script.showListPrompt('Wish', 'What creature should it be?',
+            COLOR_LIGHTGREEN, getCreatureList(), 1, true)
+        if not creatureok then return false end
+        return true, getCreatureRaceAndCaste(creatureTable[3])
+    end,
+    get_corpse_part = function(mattype, matindex)
+        return script.showListPrompt('Wish', 'What body part should it be?',
+            COLOR_LIGHTGREEN, getCreaturePartList(mattype, matindex), 1, true)
+    end,
+    get_tissue_layer = function(mattype, matindex, bodypart)
+        return script.showListPrompt('Wish', 'What tissue layer should it be?',
+            COLOR_LIGHTGREEN, getCreaturePartLayerList(mattype, matindex, bodypart), 1, true)
+    end,
+    get_creature_part_mat = function(mattype, matindex)
+        return script.showListPrompt('Wish', 'What creature material should it be?',
+            COLOR_LIGHTGREEN, getCreatureMaterialList(mattype, matindex), 1, true)
+    end,
+    get_quality = function()
+        return script.showListPrompt('Wish', 'What quality should it be?',
+            COLOR_LIGHTGREEN, qualityTable())
+    end,
+    get_description = function()
+        return script.showInputPrompt('Slab', 'What should the slab say?', COLOR_WHITE)
+    end,
+    get_count = function()
+        return script.showInputPrompt('Wish', 'How many do you want? (numbers only!)',
+            COLOR_LIGHTGREEN)
+    end,
 }
 
-local opts = utils.processArgs({...}, validArgs)
+if not dfhack.isMapLoaded() then
+    qerror('create-item needs a loaded map to work')
+end
+
+local opts = {}
+local positionals = argparse.processArgsGetopt({...}, {
+    {nil, 'startup', handler = function() opts.startup = true end},
+    {'f', 'unrestricted', handler = function() opts.unrestricted = true end},
+    {'h', 'help', handler = function() opts.help = true end},
+    {
+        'u',
+        'unit',
+        hasArg = true,
+        handler = function(arg) opts.unit = argparse.nonnegativeInt(arg, 'unit') end,
+    },
+    {
+        'c',
+        'count',
+        hasArg = true,
+        handler = function(arg) opts.count = argparse.nonnegativeInt(arg, 'count') end,
+    },
+})
+
+if positionals[1] == 'help' then opts.help = true end
+if opts.help then
+    print(dfhack.script_help())
+    return
+end
 
 if opts.startup then
     eventful.onReactionComplete.hackWishP = function(reaction, unit)
         if not reaction.code:find('DFHACK_WISH') then return end
-        hackWish(unit, {count = 1})
+        local accessors = copyall(default_accessors)
+        accessors.get_unit = function() return unit end
+        createitem.hackWish(accessors, {count = 1})
     end
     return
 end
 
-hackWish(get_creator(opts), opts)
+createitem.hackWish(default_accessors, opts)
